// src/frontend/src/shared/types/auth.types.ts
export const UserRoles = {
  SUPER_ADMIN: 'SUPER_ADMIN',
  OWNER: 'OWNER',
  ADMIN: 'ADMIN',
  TRAINER: 'TRAINER',
  NUTRITIONIST: 'NUTRITIONIST',
  CLIENT: 'CLIENT',
  EMPLOYEE: 'EMPLOYEE',
  PROFESSIONAL: 'PROFESSIONAL',
} as const;

export type UserRole = keyof typeof UserRoles;